Only update symlink when the file it wants to link to exists
git-svn-id: http://svn.dmdirc.com/trunk@3014 00569f92-eb28-0410-84fd-f71c24880f43

+++ BuildAll.sh
@@ -75,9 +75,15 @@ if [ -f $MYDIR/dist/DMDirc.jar ]; then
 	rm -Rf "${MYDIR}/installer/output/DMDirc-Setup-${FILEDATA}.*.md5"
 	cp $MYDIR/dist/DMDirc.jar /home/dmdirc/www/nightly/$FILENAME
 	
-	ln -sf $WWWDIR/nightly/$FILENAME $WWWDIR/nightly/DMDirc_latest.jar
-	ln -sf "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.run" $WWWDIR/nightly/DMDirc-Setup_latest.run
-	ln -sf "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.exe" $WWWDIR/nightly/DMDirc-Setup_latest.exe
+	if [ -e "${WWWDIR}/nightly/${FILENAME}" ]; then
+		ln -sf ${WWWDIR}/nightly/${FILENAME} $WWWDIR/nightly/DMDirc_latest.jar
+	fi;
+	if [ -e "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.run" ]; then
+		ln -sf "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.run" $WWWDIR/nightly/DMDirc-Setup_latest.run
+	fi;
+	if [ -e "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.exe" ]; then
+		ln -sf "${WWWDIR}/nightly/DMDirc-Setup-${FILEDATA}.exe" $WWWDIR/nightly/DMDirc-Setup_latest.exe
+	fi;
 	cd ${MYDIR}
 	/bin/sh $MYDIR/oblong.sh "Nightly Build" "Build Successful";
 	/bin/sh $MYDIR/DoReports.sh
